requirements.txt
Django>=1.8
django-widget-tweaks>=1.4
django-custom-user
